+++ src/Classes/StringUtil.php
@@ -90,4 +90,41 @@ class StringUtil extends \Contao\StringUtil
 
         return array_values($arrKeywords);
     }
+
+    /**
+     * Convert a number of bits to a readable filesize format
+     *
+     * @param int $bytes
+     *
+     * @return string
+     */
+    public static function formatFileSize(int $bytes): string
+    {
+        if ($bytes >= 1073741824)
+        {
+            $bytes = number_format($bytes / 1073741824, 2) . ' GB';
+        }
+        elseif ($bytes >= 1048576)
+        {
+            $bytes = number_format($bytes / 1048576, 2) . ' MB';
+        }
+        elseif ($bytes >= 1024)
+        {
+            $bytes = number_format($bytes / 1024, 2) . ' KB';
+        }
+        elseif ($bytes > 1)
+        {
+            $bytes = $bytes . ' bytes';
+        }
+        elseif ($bytes == 1)
+        {
+            $bytes = $bytes . ' byte';
+        }
+        else
+        {
+            $bytes = '0 bytes';
+        }
+
+        return $bytes;
+    }
 }