+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Extension "Utilities" for Contao Open Source CMS
 ========
 
+1.0.3 - 2024-03-28
+- Feat : ability to cancel a [DropZone](https://www.dropzone.dev/) upload
+
+1.0.2 - 2024-03-12
+- Fix missing class import
+- Feat : add file transfert is complete & path in function return when using file upload with [DropZone](https://www.dropzone.dev/)
+
+1.0.1 - 2023-09-04
+- Add a function to format a filesize
+- Add functions to manage file upload with [DropZone](https://www.dropzone.dev/)
+
+1.0.0 - 2023-08-14
+- Updated `README.md`
+- Cleaned dependencies in `composer.json`
+
 0.3.14 - 2023-08-10
 - Fix : password generation function misbehaviour